s4/torture: Add server target of OneFS
I've added a "--target=onefs" which lists expected deviation in the
OneFS SMB server implementation compared to a Windows machine.

I've added this in a generic way using a list of module specific
parameters.  This list currently only contains the absence of
SACL support but will be added to as additional server differences
are defined.

I'd liked to use this abstraction for defining the differences between
a WinXP and Win7 server as well.