s3:torture: Add a notify-bench test
This is a test that creates and deletes files in a directory as fast as the
network allows it. At the same time, it opens a filechangenotify. This test is
done to just torture handling a single directory together with the notify
infrastructure.
(cherry picked from commit 23b3fade10c5ecb5678e087e9507c527c36bdc48)