s4:torture Remove _drs_util_verify_attids() from RPC-DSSYNC
I'm satisfied that the task this test does is already done by the time
we map the incoming schema, and process the objects.  If we have the
OID mapping wrong or incomplete, we will get any errors this test
found errors there.

(And this dramaticly reduces the test time, so we can now add
RPC-DSSYNC to 'make test').

Andrew Bartlett