s3: Tune if-conditions in a very hot codepath
This looks innocent, but it is visible in a netbench run.  Due to boolean
short-circuiting we don't have to execute the conditions on the right-hand side
of the &&. So putting the less likely condition left gains a bit.
(cherry picked from commit 8cb66e2d2310567ff97e3198c37eb4c606ddc45c)